Added baseline experiments with Gemini 2.0 Flash
This includes experiments done classifying a subset of validation videos with the proprietary model: Gemini 2.0 Flash. The subset includes videos from: val.zip.001 & val.zip.002
For one-shot examples it also uses 4 extra videos extracted from: train.zip.001
The videos include examples for: A real video, a real video with fake audio, a fake video with real audio and a fake video with fake audio.
The commit also includes the metadata for the validation videos used and the results with the logs obtained from the Gemini model.
This only serves as baseline as to know how a big proprietary model's performance is when prompting this type of data.